+++ src/widget/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) mod check_box;
+pub(crate) mod multi_select;
 pub(crate) mod select;
 pub(crate) mod single_line;
 pub(crate) mod text_area;